pyproject.toml
[project]
name = "agent-skills-sdk"
version = "0.1.1"
description = "Complete SDK for integrating Agent Skills into any Python agent framework"
authors = [
    {name = "Supreeth Ravi", email = "connect@supreethravi.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["agent-skills", "ai", "llm", "agents", "langchain", "crewai", "agno"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "pathspec>=0.11.0",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
langchain = [
    "langchain>=0.1.0",
]
crewai = [
    "crewai>=0.1.0",
]
agno = [
    "agno>=0.1.0",
]
all = [
    "agent-skills-sdk[dev,langchain,crewai,agno]",
]

[project.scripts]
agent-skills = "agent_skills_sdk.cli:main"

[project.urls]
Homepage = "https://supreethravi.com"
Documentation = "https://github.com/supreeth-ravi/agentskills#readme"
Repository = "https://github.com/supreeth-ravi/agentskills"
"Bug Tracker" = "https://github.com/supreeth-ravi/agentskills/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_skills_sdk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=agent_skills_sdk --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
